refactor sshConfig changes to their own new file
